pyproject.toml
[tool.poetry]
name = "snake"
version = "0.1.0"
description = ""
authors = ["Pierrick Roger <pierrick.rogermele@cea.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.6.1"
pyyaml = "^6.0.2"
schema = "^0.7.7"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.2"
pudb = "^2024.1.3"
ipdb = "^0.13.13"
pytest = "^8.3.4"
coverage = "^7.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
snake = "snake.main:main"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212"]

[tool.mypy]
files = "."
strict = true
pretty = true
show_error_code_links = true
show_error_context = true
show_column_numbers = true
